git-svn-id: file:///home/chad/tmp/glscry/trunk/glscry@289 a68c7e1a-86db-0310-a1c4-e68a6bf46cff

+++ src/interface/GLAnalyze.py
@@ -1,25 +1,26 @@
 #!/usr/bin/env python
 #Boa:App:BoaApp
 
-from wxPython.wx import *
+import wx
 import os
 import GLAnalyzeMain
 
 modules ={u'GLAnalyzeMain': [1, 'Main frame of Application', u'GLAnalyzeMain.py']}
 
-class BoaApp(wxApp):
+class BoaApp(wx.App):
     def OnInit(self):
-        wxInitAllImageHandlers()
+        wx.InitAllImageHandlers()
         self.main = GLAnalyzeMain.create(None)
         self.main.Show()
         self.SetTopWindow(self.main)
         return True
-         
+
 def main():
     application = BoaApp(0)
     application.MainLoop()
 
 if __name__ == '__main__':
     main()
-
-os.remove("temp.png")
+    
+if os.path.exists("temp.png"):
+    os.remove("temp.png")
